==> R markdown/05_country_20240618.Rmd <==
---
title: "MAFLD"
output: html_document
date: "2024-06-17"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(bibliometrix)
library(tidyverse)
library(ggtext)
library(glue)
library(ggrepel)
library(countrycode)
library(sf)
library(rnaturalearth)
library(rnaturalearthdata)
library(rmapshaper)
library(CoordinateCleaner)
library(igraph)
library(ggraph)
library(tidygraph)
library(patchwork)
library(cowplot)
```

```{r}
# load workspace after journal analysis
load("rdata/04_journal.RData")
```

```{r}
# prepare dataframe for country analysis
df_c1 <- df_bx %>%
  mutate(C1 = case_when(!is.na(C1_w) ~ C1_w,
                        is.na(C1_w) & !is.na(C1_s) ~ C1_s,
                        is.na(C1_w) & is.na(C1_s) ~ C1_p))
nrow(df_c1)
```

```{r}
sum(is.na(df_c1$C1))
#3
```

```{r}
df_c1 %>% filter(is.na(C1))

#	10.1159/000534284
# 37535913
```

```{r}
# separate the dataframe into long format

df_c1_long <- df_c1 %>%
  mutate(C1 = str_remove(C1, "\\.;$")) %>%
  separate_longer_delim(C1, delim= ";  ") %>%
  mutate(C1 = str_remove(C1, "\\.$")) %>%
  mutate(CN = str_extract(C1, "(?<=,\\s)[^,]+$"))
```

```{r}
# check missing countries
df_c1_long %>% filter(is.na(CN)) %>% select(UI, CN)

# 10.1159/000534284		JAPAN	
# 37535913		CHINA
# 10.20960/nh.04635	SPAIN			
```

```{r}
# manually fill missing country information
df_c1_long <- df_c1_long %>%
  mutate(CN = case_when(UI == "10.1159/000534284" ~ "JAPAN",
                        UI == "37535913" ~ "CHINA",
                        UI == "10.20960/nh.04635" ~ "SPAIN",
                        .default = as.character(CN)))
```

```{r}
# format USA
df_c1_long <- df_c1_long %>%
  mutate(CN = case_when(str_detect(CN, "USA|UNITED STATES") ~ "USA",
                        CN == "CA" ~ "USA",
                        str_detect(CN, "CALIFORNIA") ~ "USA",
                        CN == "CO" ~ "USA",
                        str_detect(CN, "ICAHN SCHOOL OF MEDICINE AT MOUNT SINAI") ~ "USA",
                        .default = as.character(CN)))
```

```{r}
# format other countries
df_c1_long <- df_c1_long %>%
  mutate(CN = case_when(str_detect(CN, "ENGLAND|SCOTLAND|UK|UNITED KINGDOM") ~ "UK",
                        str_detect(CN, "CHINA|HONG KONG") ~ "CHINA",
                        str_detect(CN, "ITALY") ~ "ITALY", 
                        str_detect(CN, "SPAIN") ~ "SPAIN", 
                        str_detect(CN, "BRAZIL") ~ "BRAZIL",
                        str_detect(CN, "FRANCE") ~ "FRANCE",
                        str_detect(CN, "INDIA|NEW DELHI") ~ "INDIA",
                        str_detect(CN, "RUSSI") ~ "RUSSIA",
                        CN == "KOREA" ~ "SOUTH KOREA",
                        CN == "NORTH IRELAND" ~ "IRELAND",
                        CN == "PAK" ~ "PAKISTAN",
                        CN == "THA" ~ "THAILAND",
                        CN == "TURKIYE" ~ "TURKEY",
                        .default = as.character(CN)))
```

```{r}
# check unformatted information
df_c1_long %>%
  count(CN) %>%
  arrange(desc(n)) %>%
  View()
#97
```

```{r}
# remaining incorrect country information
# "COMPLEJO HOSPITALARIO UNIVERSITARIO DE PONTEVEDRA"
# "CORRECTED AND REPUBLISHED FROM: PANMINERVA MEDICA 2022 DECEMBER"

df_c1_long %>%
  filter(CN %in% c("COMPLEJO HOSPITALARIO UNIVERSITARIO DE PONTEVEDRA", "CORRECTED AND REPUBLISHED FROM: PANMINERVA MEDICA 2022 DECEMBER"))

# 37750860 10.23736/s0031-0808.23.04850-4 ITALY
# 36514974 10.17235/reed.2022.9325/2022 SPAIN
```

```{r}
df_c1_long %>%
  filter(UI %in% c("10.23736/s0031-0808.23.04850-4", "10.17235/reed.2022.9325/2022"))
```

```{r}
# 37750860 10.23736/s0031-0808.23.04850-4 was republished from
# 36533665 10.23736/S0031-0808.22.04730-9

df_c1_long %>% filter(DI == "10.23736/S0031-0808.22.04730-9")
df_c1_long %>% filter(PM == "36533665")

# This is not included in the dataset (no MAFLD term)
```

```{r}
# manually change the incorrect countries
df_c1_long <- df_c1_long %>%
  mutate(CN = case_when(UI == "10.23736/s0031-0808.23.04850-4" ~ "ITALY",
                        UI == "10.17235/reed.2022.9325/2022" ~ "SPAIN",
                        .default = as.character(CN)))
```

```{r}
# format country name and iso3 code using countrycode
df_c1_long <- df_c1_long %>%
  mutate(CN = countrycode(CN, origin = "country.name", destination = "country.name")) %>%
  mutate(iso_a3 = countrycode(CN, origin = "country.name", destination = "iso3c"))
```

```{r}
# count country name (full count)
df_c1_long %>%
  group_by(UI) %>%
  distinct(CN) %>%
  ungroup() %>%
  count(CN) %>%
  arrange(desc(n))
# 97 countries
```

```{r}
# count iso3 code (full count)
df_c1_long %>%
  group_by(UI) %>%
  distinct(iso_a3) %>%
  ungroup() %>%
  count(iso_a3) %>%
  arrange(desc(n))
# 97 countries
```

```{r}
#count countries and regions
df_c1_long %>%
  select(CN) %>%
  filter(!is.na(CN)) %>%
  pull(CN) %>%
  unique() %>%
  length()
# 97
```

```{r}
# publications from multiple countries
df_c1_long %>%
  group_by(UI) %>%
  distinct(CN) %>%
  filter(n() == 1) %>%
  ungroup() %>%
  count(UI) %>%
  nrow()
# 1141
```

```{r}
# the number of countries producing these publications
df_c1_long %>%
  group_by(UI) %>%
  distinct(CN) %>%
  filter(n() == 1) %>%
  ungroup() %>%
  count(CN) %>%
  nrow()
# 49
```

```{r}
# publications from multiple countries
df_c1_long %>%
  group_by(UI) %>%
  distinct(CN) %>%
  filter(n()>1) %>%
  ungroup() %>%
  count(UI)
# 328
```


```{r}
# the number of countries producing these publications
df_c1_long %>%
  group_by(UI) %>%
  distinct(CN) %>%
  filter(n()>1) %>%
  ungroup() %>%
  count(CN) %>%
  nrow()
# 96
```

```{r}
# check one missing country from multinational publications
setdiff(df_c1_long %>% distinct(CN) %>% pull(CN),
        df_c1_long %>% group_by(UI) %>% distinct(CN) %>% filter(n()>1) %>% ungroup() %>% distinct(CN) %>% pull(CN))

# "Kyrgyzstan"
```

```{r}
# plotting order of top productive countries with full counting (y axis)
cn_order_full <- df_c1_long %>%
  group_by(UI) %>%
  distinct(CN) %>%
  ungroup() %>%
  count(CN) %>%
  mutate(CN = case_when(CN == "United States" ~ "USA",
                        CN == "United Kingdom" ~ "UK",
                          .default = as.character(CN))) %>%
  arrange(desc(n)) %>%
  pull(CN)
```

```{r}
# labels for number of publications in top 20 countries
total_label_full <- df_c1_long %>%
  group_by(UI) %>%
  distinct(CN, .keep_all = TRUE) %>%
  mutate(SM = if_else(n() >1, "Multiple", "Single")) %>%
  ungroup() %>%
  count(CN) %>%
  rename(total = "n") %>%
  slice_max(total, n = 25) %>%
  mutate(CN = case_when(CN == "United States" ~ "USA",
                        CN == "United Kingdom" ~ "UK",
                          .default = as.character(CN)))
```

```{r}
# stacked bar chart for country production with full counting 
df_c1_long %>%
  group_by(UI) %>%
  distinct(CN) %>%
  mutate(SM = if_else(n()>1, "Multiple", "Single")) %>%
  ungroup() %>%
  mutate(CN = case_when(CN == "United States" ~ "USA",
                        CN == "United Kingdom" ~ "UK",
                          .default = as.character(CN))) %>%
  mutate(CN = fct_relevel(CN, rev(cn_order_full))) %>%
  filter(CN %in% cn_order_full[1:25]) %>%
  ggplot(aes(y = CN, fill = SM))+
  geom_bar()+
  geom_text(data = total_label_full,
            aes(total, CN, label = total, fill = NULL), 
            hjust = 0, vjust = 0.5, nudge_x = 2.5, size = 2.8)+
  labs(x = "Number of publications", y = "Country")+
  scale_x_continuous(breaks = seq(0, 600, 100), 
                   limits = c(0,600), 
                   expand = c(0,0))+
  scale_fill_manual(name = "Author countries", 
                    values = c("#ff7f00", "#377eb8"))+
  theme_classic()+
  theme(axis.title = element_text(size = 8, face = "bold"),
        axis.text=element_text(size=8, colour = "black"),
        axis.line = element_line(linewidth = 0.3, colour = "black"),
        axis.ticks = element_line(linewidth = 0.3, colour = "black"),
        legend.position ="inside",
        legend.position.inside = c(1, 0.1),
        legend.justification = c(1, 0),
        legend.box = "horizontal",
        legend.key.size = unit(1, "line"),
        legend.title = element_text(size = 8, face = "bold"),
        legend.text = element_text(size = 8),
        legend.title.align=0.5,
        plot.margin = margin(0,0.5,0 ,0, "cm")) 
```

```{r}
# plotting order of top productive countries with fractional counting (y axis)
cn_order_frc <- df_c1_long %>%
  group_by(UI) %>%
  distinct(CN) %>%
  count(CN) %>%
  mutate(WT = n/n()) %>%
  ungroup() %>%
  mutate(CN = case_when(CN == "United States" ~ "USA",
                        CN == "United Kingdom" ~ "UK",
                          .default = as.character(CN))) %>%
  group_by(CN) %>%
  summarise(count = sum(WT)) %>%
  arrange(desc(count)) %>%
  pull(CN)
```

```{r}
# labels for number of publications in top 25 countries
total_label_frc <- df_c1_long %>%
  group_by(UI) %>%
  distinct(CN, .keep_all = TRUE) %>%
  mutate(SM = if_else(n() >1, "Multiple", "Single")) %>%
  mutate(WT = 1/n()) %>%
  ungroup() %>%
  group_by(CN) %>%
  summarise(total = round(sum(WT))) %>%
  slice_max(total, n = 25) %>%
  mutate(CN = case_when(CN == "United States" ~ "USA",
                        CN == "United Kingdom" ~ "UK",
                          .default = as.character(CN)))
```

```{r}
# stacked bar chart for country production with fractional counting
df_c1_long %>%
  group_by(UI) %>%
  distinct(CN, .keep_all = TRUE) %>%
  mutate(SM = if_else(n() >1, "Multiple", "Single")) %>%
  mutate(WT = 1/n()) %>%
  ungroup() %>%
  group_by(CN, SM) %>%
  summarise(count = sum(WT)) %>%
  ungroup()%>%
  mutate(CN = case_when(CN == "United States" ~ "USA",
                        CN == "United Kingdom" ~ "UK",
                          .default = as.character(CN))) %>%
  mutate(CN = fct_relevel(CN, rev(cn_order_frc))) %>%
  filter(CN %in% cn_order_frc[1:25]) %>%
  ggplot(aes(count, CN, fill = SM)) +
  geom_col(width = 0.7)+
  geom_text(data = subset(total_label_frc, CN != "China"),
            aes(total, CN, label = total, fill = NULL), 
            hjust = 0, vjust = 0.5, nudge_x = 2.5, size = 2.8)+
  annotate("text", x = 487, y = "Italy", label = "487", hjust = 1, size = 2.8)+
  labs(x = "Number of publications",
       y = "Country",
       fill = "Author country")+
  scale_x_continuous(breaks = seq(0, 500, 100), 
                   limits = c(0,500), 
                   expand = c(0,0))+
  scale_fill_manual(values = c("#ff7f00", "#377eb8"))+
  theme_classic()+
  theme(axis.title = element_text(size = 8, face = "bold"),
        axis.text=element_text(size=8, colour = "black"),
        axis.line = element_line(linewidth = 0.3, colour = "black"),
        axis.ticks = element_line(linewidth = 0.3, colour = "black"),
        legend.position = "inside",
        legend.position.inside = c(1, 0.1),
        legend.justification = c(1, 0),
        legend.box = "horizontal",
        legend.key.size = unit(1, "line"),
        legend.title = element_text(size = 8, face = "bold"),
        legend.text = element_text(size = 8),
        legend.title.align=0.5,
        plot.margin = margin(0,0.5,0 ,0, "cm")) 
```

```{r}
# create world map
world <- ne_countries(scale = "medium", returnclass = "sf") %>% filter(sovereignt != "Antarctica")
```

```{r}
# check missing iso3 in world
setdiff(df_c1_long %>% distinct(iso_a3) %>% pull(iso_a3),
       world %>% distinct(iso_a3) %>% pull(iso_a3))
# "FRA" "NOR"
```

```{r}
# check invalid values in iso_a3
world %>% filter(iso_a3 == "-99") %>% View()
```

```{r}
world %>% filter(sovereignt %in% c("France", "Norway")) %>% View()
```

```{r}
# manually fix FRA and NOR
world <- world %>% mutate(iso_a3 = case_when(admin == "France" ~ "FRA",
                                    admin == "Norway" ~ "NOR",
                                    .default = as.character(iso_a3)))
```

```{r}
# load countryref containing centrid coordinates for countries
data(countryref)
```

```{r}
# get centroids from countryref
countryref <- countryref %>% 
  filter(type == "country") %>%
  mutate(centroid.lon = case_when(iso3 == "MEX" ~ capital.lon,
                                    .default = as.double(centroid.lon))) %>%
  mutate(centroid.lat = case_when(iso3 == "MEX" ~ capital.lat,
                                    .default = as.double(centroid.lat))) %>%
  filter((!is.na(centroid.lon)) & (!is.na(centroid.lat))) %>%
  distinct(iso3, .keep_all = TRUE) %>%
  select(iso3, centroid.lon, centroid.lat) %>%
  rename(iso_a3 = "iso3", 
         x = "centroid.lon",
         y ="centroid.lat")
```

```{r}
# add centroid coordinates to world sf
world <- world %>% left_join(countryref, by = join_by("iso_a3_eh" == "iso_a3"))
```

```{r}
# calculate full counts for each country
df_iso_full <- df_c1_long %>%
  group_by(UI) %>%
  distinct(iso_a3) %>%
  ungroup() %>%
  count(iso_a3) %>%
  rename(n_full = n)
```

```{r}
# calculate fractional counts for each country
df_iso_frc <- df_c1_long %>%
  group_by(UI) %>%
  distinct(iso_a3) %>%
  count(iso_a3) %>%
  mutate(WT = 1/n()) %>%
  ungroup() %>%
  group_by(iso_a3) %>%
  summarise(n_frc = sum(WT))
```

```{r}
# add full and fractional counts to world
world <- world %>% 
  left_join(df_iso_full, by = join_by(iso_a3)) %>%
  left_join(df_iso_frc, by = join_by(iso_a3))
```

```{r}
# create country label in world
world <- world %>%
  mutate(label = countrycode(iso_a3, origin = "iso3c", destination = "country.name")) %>%
  mutate(label = case_when(label == "United Kingdom" ~ "UK",
                   label == "United States" ~ "USA",
                   .default = as.character(label)))
```

```{r}
# add counts to label
world <- world %>% 
  mutate(label_full = paste0(label, "\n", "\u0028", round(n_full), "\u0029")) %>%
  mutate(label_frc = paste0(label, "\n", "\u0028", round(n_frc), "\u0029"))
```

```{r}
# get top 20 countries based on fractional counting
top_iso_frc <- df_iso_frc %>% arrange(desc(n_frc)) %>% slice_max(n_frc, n = 19) %>% pull(iso_a3)
top_iso_frc
```

```{r}
#fractional counting with label 
world %>%
  ggplot() +
  geom_sf(aes(fill = n_frc)) +
  geom_point(data = (world %>% filter(iso_a3 %in% top_iso_frc)), 
             aes(x, y),
             size = 0.5) +
  scale_fill_gradient(low= "#FFFFCC", high = "#E31A1C", na.value = "gray85",
                      limits = c(0, 500)) +
  geom_label_repel(data = subset(world, iso_a3 %in% c("USA","MEX")),
                  aes(x, y, label = label_frc),
                  nudge_x = -150 - subset(world, iso_a3 %in% c("USA", "MEX"))$x,
                  direction = "y",
                  lineheight = 1, 
                  segment.size= 0.25,
                  size = 2.6)+
  geom_label_repel(data = subset(world, iso_a3  %in% c("AUS","SGP", "IND", "TUR", "EGY")), 
                  aes(x, y, label = label_frc),
                  nudge_y = -65 - subset(world, iso_a3  %in% c("AUS", "SGP", "IND", "TUR", "EGY"))$y,
                  lineheight = 1, 
                  segment.size = 0.25,
                  size = 2.6) +
  geom_label_repel(data = subset(world, iso_a3  %in% c("BRA")), 
                  aes(x, y, label= label_frc),
                  nudge_x = -40 - subset(world, iso_a3  %in% c("BRA"))$x,
                  nudge_y = -65 - subset(world, iso_a3  %in% c("BRA"))$y,
                  lineheight = 1, 
                  segment.size = 0.25,
                  size = 2.6) +
  geom_label_repel(data = subset(world, iso_a3 %in% c("JPN", "TWN")), 
                  aes(x, y, label= label_frc),
                  nudge_x = 175 - subset(world, iso_a3 %in% c("JPN", "TWN"))$x,
                  force = 1,
                  force_pull = 0,
                  lineheight = 1, 
                  segment.size = 0.25,
                  size = 2.6)  +
  geom_label_repel(data = subset(world, iso_a3 %in% c("CHN", "ITA", "GBR", "KOR", "ESP", "DEU", "BEL", "NLD", "FRA", "POL" )), 
                  aes(x, y, label= label_frc),
                  force = 1,
                  force_pull = 0,
                  nudge_y = 110 - subset(world, iso_a3 %in% c("CHN", "ITA", "GBR", "KOR", "ESP", "DEU", "BEL", "NLD", "FRA", "POL"))$y,
                  direction = "x",
                  lineheight = 1, 
                  segment.size = 0.25,
                  size = 2.6) +
  labs( fill = "Publication counts")+
  guides(fill = guide_colourbar(ticks.colour = "black",
                                ticks.linewidth = 0.2,
                                frame.colour = "black",
                                frame.linewidth = 0.2)) +
  theme_void()+
  theme(plot.tag = element_text(size = 12, face = "bold"),
        panel.background = element_rect(fill = 	"white", colour = NA),
        legend.title = element_text(size = 8, face = "bold"),
        legend.text = element_text(size = 8),
        legend.key.width  = unit(0.75, "lines"),
        legend.key.height = unit(1, "lines"),
        legend.position = "right",
        legend.background = element_rect(fill = "white", colour = NA))
```

```{r}
# country co-occurrence network analysis
# create matrix for calculating co-occurrence matrix
# filter publications from multiple countries and convert into wide format
# row: countries CN - column: article UI
mat_c1 <- df_c1_long %>%
  group_by(UI) %>%
  distinct(CN) %>%
  filter(n()>1) %>%
  ungroup() %>%
  mutate(n = 1) %>%
  arrange(CN) %>%
  pivot_wider(names_from = UI, values_from = n) %>%
  column_to_rownames(var = "CN") %>%
  mutate(across(everything(), ~replace_na(.x, 0))) %>%
  as.matrix()
```

```{r}
# function for obtaining co-occurrence matrix
full_count <- function(A){
  U <- A %*% t(A)
  diag(U) <- 0
  return(U)
}
```

```{r}
# calculate adjacency matrix
adjm_c1_full <- full_count(mat_c1)
```

```{r}
# create igraph object from adjacency matrix
gr_c1_full <- graph_from_adjacency_matrix(adjm_c1_full, 
                                          mode = "undirected",
                                          weighted = TRUE,
                                          diag = FALSE,
                                          add.colnames = NULL,
                                          add.rownames = NA)
```

```{r}
# top co-authoring countries
gr_c1_full %>% as_long_data_frame() %>%
  mutate(across(c(from_name, to_name),
                ~ case_when(. == "United States" ~ "USA",
                            . == "United Kingdom" ~ "UK",
                            .default = as.character(.)))) %>%
  mutate(RL = paste(from_name, "-", to_name)) %>%
  arrange(desc(weight), RL) %>%
  mutate(rn = row_number()) %>%
  mutate(col = if_else(rn %in% 1:10, 1, 2)) %>%
  slice_max(weight, n = 20) %>%
  ggplot(aes(weight, fct_reorder(RL, rev(rn))))+
  geom_col(fill = "#4E98D3")+
  facet_wrap(~col, scales = "free_y")+
  labs(x ="Number of collaborations", y = "Collaborating countries")+
  scale_x_continuous(breaks = seq(0, 40, 10), 
                   limits = c(0,40), 
                   expand = c(0,0))+
  theme_classic()+
  theme(strip.background = element_blank(), strip.text = element_blank())
```

```{r}
# convert igraph object to tidygraph
tbl_gr_c1_full <- as_tbl_graph(gr_c1_full)
```

```{r}
# add iso_a3 to the nodes
tbl_gr_c1_full <- tbl_gr_c1_full %>%
  activate(nodes) %>%
  mutate(iso_a3 = countrycode(name, origin = "country.name", destination = "iso3c")) 
```

```{r}
# add weighted degree (strength) to the nodes
tbl_gr_c1_full <- tbl_gr_c1_full %>%
  activate(nodes) %>%
  mutate(degree = centrality_degree(weights = weight))
```

```{r}
# plot country ranking based on strength (weighted degree)
tbl_gr_c1_full %>%
  activate(nodes) %>%
  as_tibble() %>%
  arrange(desc(degree)) %>%
  slice_max(degree, n = 20) %>%
  ggplot(aes(degree, fct_reorder(name, degree)))+
  geom_col(fill = "#4E98D3")+
  labs(x ="Strength", y = "Country")+
  scale_x_continuous(breaks = seq(0, 400, 100), 
                   limits = c(0,400), 
                   expand = c(0,0))+
  theme_classic()
```

```{r}
# create communities using Leiden algorythem to optimase modularity
set.seed(0)
tbl_gr_c1_full <- tbl_gr_c1_full  %>%
  activate(nodes) %>%
  mutate(group = as.factor(group_leiden(weights = weight, objective_function = "modularity")))
```

```{r}
# check the number of countries in each country group
tbl_gr_c1_full   %>%
  activate(nodes) %>%
  as_tibble() %>%
  count(group)
```

```{r}
# rename group name based on group productivity
tbl_gr_c1_full <- tbl_gr_c1_full %>%
  activate(nodes) %>%
  mutate(group = case_when(group == "1" ~ "A",
                           group == "2" ~ "B",
                           group == "3" ~ "C",
                           group == "4" ~ "D",
                           group == "5" ~ "E"))
```

```{r}
# create for annotating groups
c1_group <- tbl_gr_c1_full %>%
  activate(nodes) %>%
  as_tibble() %>%
  mutate(id = row_number()) %>%
  select(id, name, iso_a3, group) 
```

```{r}
# hierachial edge bundling for country co-occurrence network
# The code is adapted from the The R Graph Gallery
# https://r-graph-gallery.com/hierarchical-edge-bundling.html

# edges is a data frame organised in hierarchical structure
# from: origin -> to: group 1-5
# from: group 1-5 -> to: each country

c1_d1 <- data.frame(from = "origin", to = paste0("group", LETTERS[1:5]))

c1_d2 <- tbl_gr_c1_full %>% 
  activate(nodes) %>%
  as_tibble() %>%
  select(group, name) %>% 
  mutate(group = paste0("group", group)) %>%
  rename(from = "group", to = "name") %>%
  arrange(from)

c1_edges <- rbind(c1_d1, c1_d2)
```

```{r}
# leaves is a vector containing all countries ("nodes" in usual graph)

all_leaves <- tbl_gr_c1_full %>% 
  activate(nodes) %>%
  as_tibble() %>%
  pull(name)
```

```{r}
# connection is a data frame of country - country connection with weights ("edges" in usual network)

c1_connect <- gr_c1_full %>% 
  as_long_data_frame() %>%
  select(from_name, to_name, weight)
```

```{r}
# vertices is a data frame of one column describing all links passing origin, group 1-5, and countries
# value can be number of publications
c1_vertices  <-  data.frame(name = unique(c(as.character(c1_edges$from), as.character(c1_edges$to))))

c1_vertices$group <- c1_edges$from[ match(c1_vertices$name, c1_edges$to ) ]
```

```{r}
#create edge_group for coloring with edge_bundle_com2
c1_vertices <- c1_vertices %>% 
  mutate(edge_group = case_when(str_detect(name, "group") ~ name, 
                                str_detect(group, "group") ~ group,
                                .default = NA))
```

```{r}
# Information for the node includes angle, horizontal adjustment, and potential flip
# calculate the angle of the labels

c1_vertices$id <- NA
c1_leaves <- which(is.na(match(c1_vertices$name, c1_edges$from)))
nleaves <- length(c1_leaves)
c1_vertices$id[ c1_leaves ] <- seq(1:nleaves)
c1_vertices$angle <- 90 - 360 * c1_vertices$id / nleaves
 
# calculate the alignment of labels: right or left
c1_vertices$hjust <- ifelse(c1_vertices$angle < -90, 1, 0)
 
# flip angle BY to make them readable
c1_vertices$angle <- ifelse(c1_vertices$angle < -90, c1_vertices$angle+180, c1_vertices$angle)
```

```{r}
# flip the label for Algeria, Croatia, Egypt
c1_vertices <- c1_vertices %>%
  mutate(angle = case_when(name == "Algeria" ~ 97.5,
                           name == "Croatia" ~ 93.75,
                           name == "Egypt"  ~ 90,
                           .default = as.numeric(angle))) %>%
  mutate(hjust = case_when(name %in% c("Algeria", "Croatia", "Egypt") ~ 1,
                           .default = as.numeric(hjust)))
```

```{r}
# Create an igraph object
heb_c1 <- graph_from_data_frame( c1_edges, vertices= c1_vertices )
```

```{r}
# the connection object must refer to the ids of the leaves
# create index of vertices
c1_vertices <- c1_vertices %>% mutate(index = row_number())
```

```{r}
# 
c1_connect <- c1_connect %>%
  left_join(c1_vertices %>% select(name, index),  by = join_by(from_name == name)) %>%
  rename(from_index = "index") %>%
  left_join(c1_vertices %>% select(name, index), by = join_by(to_name == name)) %>%
  rename(to_index = "index")
```

```{r}
# c1_d2 contains country - group relationship
c1_connect <- c1_connect %>%
  left_join(c1_d2 %>% rename(group = "from", name = "to"), by = join_by(from_name == name)) %>%
  rename(from_group = "group") %>%
  left_join(c1_d2 %>% rename(group = "from", name = "to"), by = join_by(to_name == name)) %>%
  rename(to_group = "group")
```

```{r}
# convert igraph object to tidygraph and modify country names for visualisation
tbl_heb_c1 <- heb_c1 %>%
  as_tbl_graph() %>%
  activate(nodes) %>%
  left_join(tbl_gr_c1_full %>%
              activate(nodes) %>%
              select(name, degree) %>%
              as_tibble(), by = join_by(name == name)) %>%
  mutate(name = case_when(name == "Palestinian Territories" ~ "Palestine",
                          name == "United States" ~ "USA",
                          name == "United Kingdom" ~ "UK",
                          #name == "United Arab Emirates" ~ "UAE",
                          .default = as.character(name)))
```


```{r}
#fractional counting with label 
p2_a <- world %>%
  ggplot() +
  geom_sf(aes(fill = n_frc)) +
  geom_point(data = (world %>% filter(iso_a3 %in% top_iso_frc)), 
             aes(x, y),
             size = 0.5) +
  scale_fill_gradient(low= "#FFFFCC", high = "#E31A1C", na.value = "#d9d9d9",
                      limits = c(0, 500)) +
  geom_label_repel(data = subset(world, iso_a3 %in% c("USA","MEX")),
                  aes(x, y, label = label_frc),
                  nudge_x = -150 - subset(world, iso_a3 %in% c("USA", "MEX"))$x,
                  direction = "y",
                  lineheight = 1, 
                  segment.size= 0.25,
                  size = 2.6)+
  geom_label_repel(data = subset(world, iso_a3  %in% c("AUS","SGP", "IND", "TUR", "EGY")), 
                  aes(x, y, label = label_frc),
                  nudge_y = -60 - subset(world, iso_a3  %in% c("AUS", "SGP", "IND", "TUR", "EGY"))$y,
                   force = 1,
                  force_pull = 0,
                  lineheight = 1, 
                  segment.size = 0.25,
                  size = 2.6) +
  geom_label_repel(data = subset(world, iso_a3  %in% c("BRA")), 
                  aes(x, y, label= label_frc),
                  nudge_x = -40 - subset(world, iso_a3  %in% c("BRA"))$x,
                  nudge_y = -65 - subset(world, iso_a3  %in% c("BRA"))$y,
                  lineheight = 1, 
                  segment.size = 0.25,
                  size = 2.6) +
  geom_label_repel(data = subset(world, iso_a3 %in% c("JPN", "TWN")), 
                  aes(x, y, label= label_frc),
                  nudge_x = 175 - subset(world, iso_a3 %in% c("JPN", "TWN"))$x,
                  force = 1,
                  force_pull = 0,
                  lineheight = 1, 
                  segment.size = 0.25,
                  size = 2.6)  +
  geom_label_repel(data = subset(world, iso_a3 %in% c("CHN", "ITA", "GBR", "KOR", "ESP", "DEU", "BEL", "NLD", "FRA")), 
                  aes(x, y, label= label_frc),
                  force = 1,
                  force_pull = 0,
                  nudge_y = 105 - subset(world, iso_a3 %in% c("CHN", "ITA", "GBR", "KOR", "ESP", "DEU", "BEL", "NLD", "FRA"))$y,
                  direction = "x",
                  lineheight = 1, 
                  segment.size = 0.25,
                  size = 2.6) +
  labs(tag = "",
       fill = "Publication\ncount")+
  guides(fill = guide_colourbar(ticks.colour = "black",
                                ticks.linewidth = 0.2,
                                frame.colour = "black",
                                frame.linewidth = 0.2)) +
  theme_void()+
  theme(plot.tag = element_text(size = 12, face = "bold"),
        panel.background = element_rect(fill = 	"white", colour = NA),
        legend.title = element_text(size = 8, face = "bold"),
        legend.text = element_text(size = 7),
        legend.key.width  = unit(0.75, "lines"),
        legend.key.height = unit(1, "lines"),
        legend.position = "right",
        legend.background = element_rect(fill = "white", colour = NA))
```

```{r}
# plot country co-occurrence network with hierarchical edge bundling (Fig 7a)
p2_b <- tbl_heb_c1 %>%
  ggraph(layout = 'dendrogram', circular = TRUE) + 
  geom_node_point(aes(filter = leaf, x = x*1.05, y=y*1.05, fill = group, size = degree),
                  colour = "gray30", shape = 21, stroke = 0.3)+
  scale_fill_manual(values = c("#ff7f0e","#377eb8", "#e41a1c", "#4daf4a","#984ea3"),
                    name = "",
                    labels = LETTERS[1:5],
                    guide = guide_legend(override.aes = list(size = 2.8)))+
  scale_size(range = c(1, 2.5),
             guide = "none")+
  geom_conn_bundle2(data = get_con(from = c1_connect$from_index,
                                  to = c1_connect$to_index),
                   aes(color = edge_group),
                   alpha = 0.1,
                   width = 0.3,
                   tension = 0.8,
                   lineend = "round",
                   show.legend = FALSE,
                    n = 250) +
  scale_edge_colour_manual(values = c("#ff7f0e","#377eb8", "#e41a1c", "#4daf4a","#984ea3"),
                           na.value = "#999999")+
  geom_node_text(aes(x = x*1.1, y=y*1.1, filter = leaf, label=name, angle = angle, hjust=hjust),
                 size=2.6,
                 alpha=1) +
  coord_fixed() +
  expand_limits(x = c(-1.4, 3.3), y = c(-1.4, 1.4))+
  labs(tag = "b")+
  theme_void() +
  theme(plot.tag = element_text(size = 12, face = "bold"),
        plot.margin=unit(c(0,0,0,0),"cm"),
        legend.title = element_blank(),
        legend.text = element_text(size = 8),
        legend.direction = "horizontal")
```

```{r}
# extract legend from p2_b
p2_b_legend <- get_legend(p2_b)
```

```{r}
# remove legend from p3_b
p2_b <- p2_b + theme(legend.position = "none")
```

```{r}
# world map couloured by country group in the country co-occurrence netowork (Fig 7b)
p2_b2 <- world %>%
  left_join(c1_group, by = join_by(iso_a3)) %>%
  ms_filter_islands(min_area = 1.1e10) %>%
  ggplot() +
  geom_sf(aes(fill = group), alpha = 0.5) +
  scale_fill_manual(values = c("#ff7f0e","#377eb8", "#e41a1c", "#4daf4a","#984ea3"),
                    na.value = "#A9A9A9", guide = "none") +
 labs(title = "Country group")+
  theme_void()+
  theme(plot.title = element_text(size = 8, face = "bold", hjust = 0.5))
```

```{r}
# fractional counting for the number of publications in each country group (Fig 7c)
p2_c <- df_c1_long %>%
  left_join(c1_group, by = join_by(iso_a3)) %>%
  filter(!is.na(group)) %>%
  group_by(UI) %>%
  distinct(iso_a3, .keep_all = TRUE) %>%
  filter(n()>1) %>%
  mutate(WT = 1/n()) %>%
  ungroup() %>%
  group_by(group) %>%
  summarise(count = sum(WT)) %>%
  ggplot(aes(group, count, fill = group))+
  geom_col(width = 0.7)+
  geom_text(aes(label = round(count)),
            size = 2.8,
            vjust = 0,
            nudge_y = 5) +
  scale_fill_manual(values = c("#ff7f0e","#377eb8", "#e41a1c", "#4daf4a","#984ea3"),
                    guide = "none")+
  scale_y_continuous(breaks = seq(0, 200, 50), 
                   limits = c(0,200), 
                   expand = c(0,0))+
  labs(x = "Country group", y = "Publication count")+
 theme_classic()+
  theme(plot.tag = element_text(size = 12, face = "bold"),
        axis.title = element_text(size = 8, face = "bold", colour = "black"),
        axis.text = element_text(size = 8, colour = "black"),
        axis.line = element_line(linewidth = 0.3, colour = "black"),
        axis.ticks = element_line(linewidth = 0.3, colour = "black"),
        plot.margin=unit(c(0,0,0,5),"mm"))

```

```{r}
# create a dataframe for plotting publications from large number of countries
df_c1_multi <- df_c1_long %>%
  group_by(UI) %>%
  distinct(CN, .keep_all = TRUE) %>%
  mutate(CN_n = n()) %>%
  filter(CN_n > 1) %>%
  ungroup() %>%
  distinct(UI, .keep_all = TRUE) %>%
  mutate(PN2 = case_when(PM == "32278004" ~ "Eslam M (2020) [a]",
                   PM == "33006093" ~ "Eslam M (2020) [b]",
                   PM == "34364544" ~ "Eslam M (2020) [c]",
                   PM == "33031758" ~ "Shiha G (2021) [a]",
                   PM == "33181119" ~ "Shiha G (2021) [b]",
                   .default = as.character(PN2))) %>%
  mutate(TI = str_remove(TI, "COMMENTARY: ")) %>%
  mutate(TI = str_to_sentence(TI)) %>%
  mutate(TI = str_replace(TI, "mafld", "MAFLD")) %>%
  mutate(TI = str_replace(TI, "Mafld", "MAFLD")) %>%
  mutate(TI = str_replace(TI, "nafld", "NAFLD")) %>%
  mutate(TI = str_replace(TI, "cvd", "CVD")) %>%
  mutate(TI = str_replace(TI, "ckd", "CKD")) %>%
  mutate(TI = str_replace(TI, "delphi", "Delphi")) %>%
  mutate(TI = str_replace(TI, "The asian pacific association for the study of the liver", "The Asian Pacific Association for the Study of the Liver")) %>%
  mutate(TI = str_replace(TI, "middle east and north africa", "Middle East and north Africa")) %>%
  mutate(TI = str_replace(TI, "saharan africa", "Saharan Africa")) %>%
  mutate(TI = str_replace(TI, "The latin american association for the study of the liver", "The Latin American Association for the Study of the Liver")) %>%
  mutate(TI = str_replace(TI, "aleh", "ALEH")) %>%
  mutate(TI = str_replace(TI, "Wgo", "WGO")) %>%
  mutate(TI = str_replace(TI, "covid-19", "COVID-19"))
```

```{r}
# publications from high number of countries (Fig 8d)
p2_d <- df_c1_multi %>%
  filter(CN_n > 10) %>%
  arrange(CN_n, desc(PY), desc(TI)) %>%
  mutate(order = row_number()) %>%
  ggplot(aes(CN_n, fct_reorder(PN2, order)))+
  geom_col(fill = "#f781bf", alpha = 0.5, width = 0.7)+
  geom_text(aes(x = 0.5, label = TI),
            hjust = 0, vjust = 0.5,
            size = 2.8)+
   scale_x_continuous(breaks = seq(0, 30, 5), 
                   limits = c(0,30), 
                   expand = c(0,0))+
  labs(x = "Number of author countries", y = "Publication")+
  theme_classic()+
  theme(plot.tag = element_text(size = 12,face="bold"),
        axis.text=element_text(size=8, colour = "black"),
        axis.title = element_text(size = 8, face = "bold"),
        axis.line = element_line(linewidth = 0.3, colour = "black"),
        axis.ticks = element_line(linewidth = 0.3, colour = "black"),
        plot.margin = margin(0,0.5,0,0, "cm")) 
```

```{r}
# plot the figure for country analysis (Fig 6)
p2 <- free(p2_a) +
  free(p2_b) +
  inset_element(p2_b2, 0.45, 0.55, 1, 1, align_to = 'full',
                ignore_tag = TRUE)+
  inset_element(p2_b_legend, 0.6, 0.55, 1, 0.575, 
                align_to = 'full', ignore_tag = TRUE, on_top = FALSE)+
  inset_element(p2_c, 0.625, 0, 1, 0.5, align_to = 'full')+
  free(p2_d)  +
  plot_layout(heights = c(3, 4, 2))+
  plot_annotation(tag_levels = 'a')

ggsave("figure/Fig6.pdf", p2, width = 174, height = 234, unit = "mm")
```

```{r}
# end of country analysis
save.image(file = "rdata/05_country.RData")
```
